server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "steady-path-team"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openai (>=1.96.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
